test(hogql): fold parser regressions into parser_test_factory suite
Moves all 103 tests from `test_parser_regressions.py` into the shared
`_test_parser.py` factory, so each one runs against `cpp-json`,
`rust-json`, and `python` via the existing backend-spawning subclasses
instead of looping internally. The factory pattern handles
backend-iteration, and inline `for backend in _BACKENDS:` /
`for backend in ("rust-json", "python"):` loops are dropped.

Two python-specific divergences are surfaced as explicit skips:
  - `test_settings_and_top_clause_error_class`: python lacks the JSON
    wrapper layer that converts `NotImplementedError` to
    `ExposedHogQLError`, so it leaks the raw class.
  - `test_window_frame_non_int_bound_keeps_constant`: python visitor
    diverges from cpp on non-int frame bounds (separate cpp-vs-python
    issue).

Adds a `@no_memory_leak_check` decorator on `MemoryLeakTestMixin`.
Tests that assert a parser raises (62 of the 103 regressions) trip
the mixin's incremental-memory ratio check on cpp's verbose ANTLR
error strings even though nothing actually leaks — the decorator
short-circuits the priming/measure loop and runs the method once.

Net: -45 lines (regressions file deleted, factory-level wiring shared).
1124 tests pass across all 3 backends; 3 skipped (the 2 python-only
divergences above + the existing rust position-emission skip).
